requirements.txt
google-generativeai==0.8.5
python-dotenv==1.0.1
ruff==0.13.0